Added Alarm to trigger Nest to come on
Use Case: When I set Nest to away when we are away from the house, I
want to be able to set a date and time for the Nest to be set to Home.
So, when we get back home, the house is at a comfortable temperature.
There have been multiple times where I think about setting the Nest to
home when we start traveling, but forget to actually set it to home at
the correct time.